Reviews are being created and persist in the db. Need to make page to show them in browser
